requirements.txt
chardet==5.0.0
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Excel==0.0.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.1
openpyxl==3.0.10
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pymongo==4.3.2
python-dateutil==2.8.2
six==1.16.0
texttable==1.6.4
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.10.0
chardet==5.0.0
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Excel==0.0.7
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.1
openpyxl==3.0.10
passlib==1.7.4
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pymongo==4.3.2
python-dateutil==2.8.2
six==1.16.0
texttable==1.6.4
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.10.0
chardet==5.0.0
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Excel==0.0.7
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.1
openpyxl==3.0.10
passlib==1.7.4
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pymongo==4.3.2
PyPDF2==2.11.1
python-dateutil==2.8.2
six==1.16.0
texttable==1.6.4
typing_extensions==4.4.0
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.10.0
